refactor(discord): bootstrap の store/context レイヤ生成を layers.ts へ分離する
